POC for styling app via ref
This is an interesting experiment. It shows that the style of the circle
can be set via CSS variables being changed on a ref of the component.

The way that the CSS is currently written means that this approach is
not compatible with the .cf-theme configuration setup, which might be a
hint that this isn't the best approach.

This supports the idea that CSS file based config and prop-based config
aren't compatible. One potential solution may be to allow the user to
pass two props: "configureViaCSS" and "customClassName". if
configureViaCSS is true, then we add customClassName? customClassName :
'cf-wedge' and that would enable css-level config. Otherwise, we would
only do the component's default style, and a user could do a
document.documentElement.style.setProperty to set theme variables (or we
could do that in the component when consuming a theme object).

definitely worth further investigation.